p4i integration Integrate 171118 from 4.0 to main: Fixes:    compilation of embedded on amd64 Task:     none RevBy:    Marius AutoTest: none Details:
[git-p4: depot-paths = "//depot/qt/main/": change = 171119]

+++ src/corelib/kernel/qsharedmemory_p.cpp
@@ -109,7 +109,7 @@ bool QSharedMemory::attach ()
     shmId = shmget (key, shmSize, 0);
 
   shmBase = shmat (shmId, 0, 0);
-  if ((int) shmBase == -1 || shmBase == 0)
+  if ((long) shmBase == -1 || shmBase == 0)
     return false;
   else
     return true;
