+++ src/gui/embedded/qwsregionmanager_qws.cpp
@@ -269,7 +269,7 @@ bool QWSRegionManager::attach(const QString &filename)
             data = (unsigned char *)shmat(shmId, 0, 0);
     }
 
-    return (shmId != -1 && (int)data != -1);
+    return (shmId != -1 && (long)data != -1);
 #else
         int dataSize = sizeof(QWSRegionHeader)                // header
                     + sizeof(QWSRegionIndex) * QT_MAX_REGIONS // + index
